requirements.txt
# Core dependencies
python-dotenv
requests

# Azure services
azure-identity
azure-storage-blob
azure-data-tables  # For newsletter subscriber management
azure-ai-textanalytics
azure-search-documents
azure-communication-email  # Optional: For email newsletter delivery

# Web scraping
beautifulsoup4
feedparser

# Streamlit dashboard
streamlit
plotly
pandas
numpy
seaborn
matplotlib
scipy
wordcloud
python-dateutil

# AI/ML
openai

# Development only (not needed for production)
# ipykernel
